+++ src/natcap/invest/coastal_blue_carbon/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Global Variables
-NODATA_FLOAT = -16777216
-NODATA_INT = -9999
-HA_PER_M2 = 0.0001