+++ tests/test_coastal_blue_carbon.py
@@ -11,9 +11,6 @@ from osgeo import gdal
 from pygeoprocessing import geoprocessing as geoprocess
 import pygeoprocessing.testing as pygeotest
 
-from natcap.invest.coastal_blue_carbon import coastal_blue_carbon as cbc
-from natcap.invest.coastal_blue_carbon import io, preprocessor
-
 pp = pprint.PrettyPrinter(indent=4)
 
 
@@ -39,7 +36,8 @@ carbon_pool_initial_list = \
      ['3', 'Z', '20', '20', '0.5']]
 
 carbon_pool_transient_list = \
-    [['code', 'lulc-class', 'pool', 'half-life', 'med-impact-dist', 'yearly_accumulation'],
+    [['code', 'lulc-class', 'pool', 'half-life', 'med-impact-dist',
+        'yearly_accumulation'],
      ['0', 'N', 'biomass', '0', '0', '0'],
      ['0', 'N', 'soil', '0', '0', '0'],
      ['1', 'X', 'biomass', '1', '0.5', '1'],
@@ -51,6 +49,7 @@ carbon_pool_transient_list = \
 
 
 def create_table(uri, rows_list):
+    """Create csv file from list of lists."""
     with open(uri, 'w') as f:
         writer = csv.writer(f)
         writer.writerows(rows_list)
@@ -62,7 +61,7 @@ class TestIO(unittest.TestCase):
     """Test io library functions."""
 
     def setUp(self):
-        band_matrices = [np.ones((2,2))]
+        band_matrices = [np.ones((2, 2))]
         srs = pygeotest.sampledata.SRS_WILLAMETTE
 
         path = os.path.dirname(os.path.realpath(__file__))
@@ -71,22 +70,28 @@ class TestIO(unittest.TestCase):
             shutil.rmtree(workspace)
         os.mkdir(workspace)
         lulc_lookup_uri = create_table(os.path.join(workspace,
-            'lulc_lookup.csv'), lulc_lookup_list)
-        lulc_transition_matrix_uri = create_table(os.path.join(workspace,
-            'lulc_transition_matrix.csv'), lulc_transition_matrix_list)
-        carbon_pool_initial_uri = create_table(os.path.join(workspace,
-            'carbon_pool_initial.csv'), carbon_pool_initial_list)
-        carbon_pool_transient_uri = create_table(os.path.join(workspace,
-            'carbon_pool_transient.csv'), carbon_pool_transient_list)
+                                       'lulc_lookup.csv'), lulc_lookup_list)
+        lulc_transition_matrix_uri = create_table(
+            os.path.join(workspace, 'lulc_transition_matrix.csv'),
+            lulc_transition_matrix_list)
+        carbon_pool_initial_uri = create_table(
+            os.path.join(workspace, 'carbon_pool_initial.csv'),
+            carbon_pool_initial_list)
+        carbon_pool_transient_uri = create_table(
+            os.path.join(workspace, 'carbon_pool_transient.csv'),
+            carbon_pool_transient_list)
         raster_0_uri = pygeotest.create_raster_on_disk(
             band_matrices, srs.origin, srs.projection, -1, srs.pixel_size(100),
-            datatype=gdal.GDT_Int32, filename=os.path.join(workspace, 'raster_0.tif'))
+            datatype=gdal.GDT_Int32, filename=os.path.join(
+                workspace, 'raster_0.tif'))
         raster_1_uri = pygeotest.create_raster_on_disk(
             band_matrices, srs.origin, srs.projection, -1, srs.pixel_size(100),
-            datatype=gdal.GDT_Int32, filename=os.path.join(workspace, 'raster_1.tif'))
+            datatype=gdal.GDT_Int32, filename=os.path.join(
+                workspace, 'raster_1.tif'))
         raster_2_uri = pygeotest.create_raster_on_disk(
             band_matrices, srs.origin, srs.projection, -1, srs.pixel_size(100),
-            datatype=gdal.GDT_Int32, filename=os.path.join(workspace, 'raster_2.tif'))
+            datatype=gdal.GDT_Int32, filename=os.path.join(
+                workspace, 'raster_2.tif'))
         lulc_baseline_map_uri = raster_0_uri
         lulc_transition_maps_list = [raster_1_uri, raster_2_uri]
 
@@ -110,8 +115,13 @@ class TestIO(unittest.TestCase):
         }
 
     def test_get_inputs(self):
+        from natcap.invest.coastal_blue_carbon import io
         d = io.get_inputs(self.args)
         self.assertTrue(d['lulc_to_Hb'][0] == 0.0)
+        self.assertTrue(d['lulc_to_Hb'][1] == 1.0)
+        self.assertTrue(len(d['price_t']) == 11)
+        self.assertTrue(len(d['snapshot_years']) == 3)
+        self.assertTrue(len(d['transition_years']) == 2)
 
     def tearDown(self):
         shutil.rmtree(self.args['workspace_dir'])
@@ -122,7 +132,7 @@ class TestModel(unittest.TestCase):
     """Test main model functions."""
 
     def setUp(self):
-        band_matrices = [np.ones((2,2))]
+        band_matrices = [np.ones((2, 2))]
         srs = pygeotest.sampledata.SRS_WILLAMETTE
 
         path = os.path.dirname(os.path.realpath(__file__))
@@ -130,23 +140,30 @@ class TestModel(unittest.TestCase):
         if os.path.exists(workspace):
             shutil.rmtree(workspace)
         os.mkdir(workspace)
-        lulc_lookup_uri = create_table(os.path.join(workspace,
-            'lulc_lookup.csv'), lulc_lookup_list)
-        lulc_transition_matrix_uri = create_table(os.path.join(workspace,
-            'lulc_transition_matrix.csv'), lulc_transition_matrix_list)
-        carbon_pool_initial_uri = create_table(os.path.join(workspace,
-            'carbon_pool_initial.csv'), carbon_pool_initial_list)
-        carbon_pool_transient_uri = create_table(os.path.join(workspace,
-            'carbon_pool_transient.csv'), carbon_pool_transient_list)
+        lulc_lookup_uri = create_table(
+            os.path.join(workspace, 'lulc_lookup.csv'),
+            lulc_lookup_list)
+        lulc_transition_matrix_uri = create_table(
+            os.path.join(workspace, 'lulc_transition_matrix.csv'),
+            lulc_transition_matrix_list)
+        carbon_pool_initial_uri = create_table(
+            os.path.join(workspace, 'carbon_pool_initial.csv'),
+            carbon_pool_initial_list)
+        carbon_pool_transient_uri = create_table(
+            os.path.join(workspace, 'carbon_pool_transient.csv'),
+            carbon_pool_transient_list)
         raster_0_uri = pygeotest.create_raster_on_disk(
             band_matrices, srs.origin, srs.projection, -1, srs.pixel_size(100),
-            datatype=gdal.GDT_Int32, filename=os.path.join(workspace, 'raster_0.tif'))
+            datatype=gdal.GDT_Int32, filename=os.path.join(
+                workspace, 'raster_0.tif'))
         raster_1_uri = pygeotest.create_raster_on_disk(
             band_matrices, srs.origin, srs.projection, -1, srs.pixel_size(100),
-            datatype=gdal.GDT_Int32, filename=os.path.join(workspace, 'raster_1.tif'))
+            datatype=gdal.GDT_Int32, filename=os.path.join(
+                workspace, 'raster_1.tif'))
         raster_2_uri = pygeotest.create_raster_on_disk(
             band_matrices, srs.origin, srs.projection, -1, srs.pixel_size(100),
-            datatype=gdal.GDT_Int32, filename=os.path.join(workspace, 'raster_2.tif'))
+            datatype=gdal.GDT_Int32, filename=os.path.join(
+                workspace, 'raster_2.tif'))
         lulc_baseline_map_uri = raster_0_uri
         lulc_transition_maps_list = [raster_1_uri, raster_2_uri]
 
@@ -171,6 +188,8 @@ class TestModel(unittest.TestCase):
 
     def test_model_run(self):
         """Test main model 'run' function."""
+        from natcap.invest.coastal_blue_carbon \
+            import coastal_blue_carbon as cbc
         cbc.execute(self.args)
         output_raster = os.path.join(
             os.path.split(os.path.realpath(__file__))[0],
@@ -190,8 +209,8 @@ class TestPreprocessor(unittest.TestCase):
     """Test preprocessor library functions."""
 
     def setUp(self):
-        band_matrices_zeros = [np.zeros((2,2))]
-        band_matrices_ones = [np.ones((2,2))]
+        band_matrices_zeros = [np.zeros((2, 2))]
+        band_matrices_ones = [np.ones((2, 2))]
         srs = pygeotest.sampledata.SRS_WILLAMETTE
 
         path = os.path.dirname(os.path.realpath(__file__))
@@ -200,21 +219,26 @@ class TestPreprocessor(unittest.TestCase):
             shutil.rmtree(workspace)
         os.mkdir(workspace)
 
-        lulc_lookup_uri = create_table(os.path.join(workspace,
-            'lulc_lookup.csv'), lulc_lookup_list)
+        lulc_lookup_uri = create_table(
+            os.path.join(workspace, 'lulc_lookup.csv'),
+            lulc_lookup_list)
 
         raster_0_uri = pygeotest.create_raster_on_disk(
-            band_matrices_ones, srs.origin, srs.projection, -1, srs.pixel_size(100),
-            datatype=gdal.GDT_Int32, filename=os.path.join(workspace, 'raster_0.tif'))
+            band_matrices_ones, srs.origin, srs.projection, -1,
+            srs.pixel_size(100), datatype=gdal.GDT_Int32,
+            filename=os.path.join(workspace, 'raster_0.tif'))
         raster_1_uri = pygeotest.create_raster_on_disk(
-            band_matrices_ones, srs.origin, srs.projection, -1, srs.pixel_size(100),
-            datatype=gdal.GDT_Int32, filename=os.path.join(workspace, 'raster_1.tif'))
+            band_matrices_ones, srs.origin, srs.projection, -1,
+            srs.pixel_size(100), datatype=gdal.GDT_Int32,
+            filename=os.path.join(workspace, 'raster_1.tif'))
         raster_2_uri = pygeotest.create_raster_on_disk(
-            band_matrices_ones, srs.origin, srs.projection, -1, srs.pixel_size(100),
-            datatype=gdal.GDT_Int32, filename=os.path.join(workspace, 'raster_2.tif'))
+            band_matrices_ones, srs.origin, srs.projection, -1,
+            srs.pixel_size(100), datatype=gdal.GDT_Int32,
+            filename=os.path.join(workspace, 'raster_2.tif'))
         raster_3_uri = pygeotest.create_raster_on_disk(
-            band_matrices_zeros, srs.origin, srs.projection, -1, srs.pixel_size(100),
-            datatype=gdal.GDT_Int32, filename=os.path.join(workspace, 'raster_3.tif'))
+            band_matrices_zeros, srs.origin, srs.projection, -1,
+            srs.pixel_size(100), datatype=gdal.GDT_Int32,
+            filename=os.path.join(workspace, 'raster_3.tif'))
 
         self.args = {
             'workspace_dir': workspace,
@@ -231,17 +255,23 @@ class TestPreprocessor(unittest.TestCase):
         }
 
     def test_preprocessor(self):
-        path = os.path.dirname(os.path.realpath(__file__))
+        from natcap.invest.coastal_blue_carbon import preprocessor
         preprocessor.execute(self.args)
-        trans_csv = os.path.join(path, 'workspace/outputs_preprocessor/transitions_test.csv')
+        trans_csv = os.path.join(
+            self.args['workspace_dir'],
+            'outputs_preprocessor',
+            'transitions_test.csv')
         with open(trans_csv, 'r') as f:
             lines = f.readlines()
         self.assertTrue(lines[2][:].startswith('Z,,accum'))
 
     def test_preprocessor_2(self):
-        path = os.path.dirname(os.path.realpath(__file__))
+        from natcap.invest.coastal_blue_carbon import preprocessor
         preprocessor.execute(self.args2)
-        trans_csv = os.path.join(path, 'workspace/outputs_preprocessor/transitions_test.csv')
+        trans_csv = os.path.join(
+            self.args2['workspace_dir'],
+            'outputs_preprocessor',
+            'transitions_test.csv')
         with open(trans_csv, 'r') as f:
             lines = f.readlines()
         self.assertTrue(lines[2][:].startswith('Z,disturb,accum'))