+++ HISTORY.rst
@@ -13,6 +13,7 @@
 * Refactored all the InVEST model user interfaces so that Workspace defaults to the user's home "Documents" directory.
 * Fixed an HRA bug where stessors with a buffer of zero were being buffered by 1 pixel
 * HRA enhancement which creates a common raster to burn all input shapefiles onto, ensuring consistent alignment.
+* Refactored Coastal Blue Carbon model for greater speed, maintainability and clearer documentation.
 
 3.2.1
 =====